Revert "Implement design"
