+++ feature/links/hooks/useLinkInput.ts
@@ -3,36 +3,30 @@ import Toast from "react-native-toast-message";
 import { mutate } from "swr";
 
 import { addLinkAndUser } from "../service/linkServices";
+import { useOGData } from "./useOGData";
 
 export const useLinkInput = (userId: string | undefined) => {
+  const [url, setUrl] = useState<string>("");
   const [isSubmitting, setIsSubmitting] = useState(false);
+  const { ogData, isLoading, isError } = useOGData(url);
 
-  const handleAddLink = async (url: string): Promise<boolean> => {
-    if (!userId || !url) return false;
+  const handleAddLink = async (): Promise<void> => {
+    if (!userId || !url) return;
 
     try {
       setIsSubmitting(true);
-      const data = await addLinkAndUser(url, userId);
+      const { data } = await addLinkAndUser(url, userId);
       await mutate(
-        (key) =>
-          (typeof key === "string" && key.startsWith("links-")) ||
-          (Array.isArray(key) &&
-            typeof key[0] === "string" &&
-            key[0].startsWith("links-")),
+        (key) => typeof key === "string" && key.startsWith("links-"),
       );
-
+      setUrl("");
       Toast.show({
-        text1:
-          data === "registered"
-            ? "Success, Link added"
-            : "This link is already registered",
+        text1: data === "registered" ? "Success" : "Already registered",
         type: data === "registered" ? "success" : "info",
         position: "top",
         topOffset: 70,
         visibilityTime: 3000,
       });
-
-      return data !== "";
     } catch (error: Error | unknown) {
       Toast.show({
         text1: error instanceof Error ? error.message : "Failed to add link",
@@ -41,14 +35,18 @@ export const useLinkInput = (userId: string | undefined) => {
         topOffset: 70,
         visibilityTime: 3000,
       });
-      return false;
     } finally {
       setIsSubmitting(false);
     }
   };
 
   return {
+    url,
+    setUrl,
     isSubmitting,
+    ogData,
+    isLoading,
+    isError,
     handleAddLink,
   };
 };
