+++ feature/links/views/LinkInputView.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import { View } from "react-native";
+import { useEffect, useRef } from "react";
+import { View, type TextInput } from "react-native";
 
 import { LinkIcon } from "@/components/icons/LinkIcon";
 import { HalfModal } from "@/components/layout/HalfModal";
@@ -10,45 +10,37 @@ import { LinkInputActions } from "../components/actions/LinkInputActions";
 import { LinkInputForm } from "../components/forms/linkInputForm";
 import { LinkPreview } from "../components/preview/LinkPreview";
 import { useLinkInput } from "../hooks/useLinkInput";
-import { useOGData } from "../hooks/useOGData";
 
 export const LinkInputView = () => {
   const { isOpen, closeModal } = useLinkInputModal();
   const { session } = useSessionContext();
-  const [url, setUrl] = useState("");
-  const { isSubmitting, handleAddLink } = useLinkInput(session?.user?.id);
-  const { ogData, isLoading, isError } = useOGData(url);
-
-  const clearForm = () => {
-    setUrl("");
-  };
-
-  const handleClose = () => {
-    clearForm();
-    closeModal();
-  };
+  const inputRef = useRef<TextInput>(null);
+  const {
+    url,
+    setUrl,
+    isSubmitting,
+    ogData,
+    isLoading,
+    isError,
+    handleAddLink,
+  } = useLinkInput(session?.user?.id);
 
   useEffect(() => {
     if (isOpen) {
-      clearForm();
+      setTimeout(() => {
+        inputRef.current?.focus();
+      }, 100);
     }
   }, [isOpen]);
 
-  const handleSubmit = async () => {
-    const success = await handleAddLink(url);
-    if (success) {
-      setUrl("");
-    }
-  };
-
   return (
-    <HalfModal isOpen={isOpen} onClose={handleClose}>
+    <HalfModal isOpen={isOpen} onClose={closeModal}>
       <View className="flex-1 gap-4">
         <View className="flex-row items-center justify-start gap-2">
           <LinkIcon size={16} color="#FA4714" />
           <Title title="Add a Link" />
         </View>
-        <LinkInputForm url={url} onUrlChange={setUrl} />
+        <LinkInputForm onUrlChange={setUrl} />
         <LinkPreview
           url={url}
           ogData={ogData ?? null}
@@ -56,8 +48,8 @@ export const LinkInputView = () => {
           isError={isError}
         />
         <LinkInputActions
-          onCancel={handleClose}
-          onAdd={handleSubmit}
+          onCancel={closeModal}
+          onAdd={handleAddLink}
           isLoading={isLoading}
           isError={isError}
           isSubmitting={isSubmitting}
